+++ bootique/src/main/java/io/bootique/env/Environment.java
@@ -10,13 +10,6 @@ public interface Environment {
 
     String FRAMEWORK_PROPERTIES_PREFIX = "bq";
 
-    /**
-     * @since 0.17
-     * @deprecated since 0.24 as BQ_ variables are deprecated.
-     */
-    @Deprecated
-    String FRAMEWORK_VARIABLES_PREFIX = "BQ_";
-
     String getProperty(String name);
 
     /**
@@ -38,41 +31,5 @@ public interface Environment {
      */
     default Map<String, String> frameworkProperties() {
         return subproperties(FRAMEWORK_PROPERTIES_PREFIX);
-    }
-
-    /**
-     * Returns a value of the environment variable with a given name.
-     *
-     * @param name environment variable name.
-     * @return a value of the environment variable with a given name.
-     * @since 0.17
-     * @deprecated since 0.24 BQ_ variables are deprecated and we may no longer need to track variables inside the stack.
-     */
-    @Deprecated
-    String getVariable(String name);
-
-    /**
-     * Returns a map of environment variables that start with a prefix. Prefix
-     * is stripped from the returned names.
-     *
-     * @param prefix a prefix to qualify variables with.
-     * @return a map of environment variables that start with a prefix.
-     * @since 0.17
-     * @deprecated since 0.24 BQ_ variables are deprecated and we may no longer need to track variables inside the stack.
-     */
-    @Deprecated
-    Map<String, String> variables(String prefix);
-
-    /**
-     * Returns a map of all variables that start with "BQ_" prefix. Prefix is
-     * stripped from the returned names.
-     *
-     * @return a map of all variables that start with "BQ_" prefix.
-     * @since 0.17
-     * @deprecated since 0.24 as BQ_ variables are deprecated.
-     */
-    @Deprecated
-    default Map<String, String> frameworkVariables() {
-        return variables(FRAMEWORK_VARIABLES_PREFIX);
     }
 }