+++ bootique/src/test/java/io/bootique/JsonNodeConfigurationFactoryProviderTest.java
@@ -29,11 +29,6 @@ public class JsonNodeConfigurationFactoryProviderTest {
     }
 
     private static ConfigurationFactory factory(Map<String, String> propertyOverrides, String... yaml) {
-        return factory(propertyOverrides, Collections.emptyMap(), yaml);
-    }
-
-    private static ConfigurationFactory factory(Map<String, String> propertyOverrides, Map<String, String> varOverrides,
-                                                String... yaml) {
         ConfigurationSource mockSource = mock(ConfigurationSource.class);
 
         when(mockSource.get()).thenReturn(asList(yaml).stream().map(s -> {
@@ -54,7 +49,6 @@ public class JsonNodeConfigurationFactoryProviderTest {
 
         Environment mockEnvironment = mock(Environment.class);
         when(mockEnvironment.frameworkProperties()).thenReturn(propertyOverrides);
-        when(mockEnvironment.frameworkVariables()).thenReturn(varOverrides);
 
         return factory(mockSource, mockEnvironment);
     }
@@ -126,30 +120,6 @@ public class JsonNodeConfigurationFactoryProviderTest {
         assertNotNull(b1);
         assertEquals("SS", b1.s);
         assertEquals(55, b1.i);
-    }
-
-    @Test
-    public void testGet_SingleConfig_PropOverride_VarOverride_Nested() {
-
-        Map<String, String> propOverrides = new HashMap<String, String>() {
-
-            {
-                put("b1.s", "SS");
-                put("b1.i", "55");
-            }
-        };
-
-        Map<String, String> varOverrides = new HashMap<String, String>() {
-
-            {
-                put("B1_I", "58");
-            }
-        };
-
-        Bean1 b1 = factory(propOverrides, varOverrides, "b1:\n  s: replace_me\n  i: 6").config(Bean1.class, "b1");
-        assertNotNull(b1);
-        assertEquals("SS", b1.s);
-        assertEquals(58, b1.i);
     }
 
     public static class Bean1 {