+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.26
 
 * #188 Update Guice to 4.2.0
+* #214 Cleaning up APIs deprecated since <= 0.25
 
 ## 0.25